Refactor pubsub base classes to package init
