--- pubsub/publisher.py
@@ -1,27 +1,0 @@
-import logging
-from abc import ABC, abstractmethod
-from concurrent.futures import ThreadPoolExecutor
-from typing import List
-
-from commons.data import Response
-from pubsub.connection import Connection
-
-logger = logging.getLogger('root')
-
-
-class Publisher(ABC):
-
-    def __init__(self, connection: Connection, publisher_thread_pool: ThreadPoolExecutor):
-        self.connection = connection
-        self.publisher_thread_pool = publisher_thread_pool
-        if connection.is_connected():
-            logger.debug(f'Using the connection {connection.__class__.__name__} for publishing.')
-        else:
-            raise ConnectionError(f'Connection failed: {connection.__class__.__name__}')
-
-    @abstractmethod
-    def _publish(self, msgs: List[Response], topic: str, schema_name=None) -> None:
-        pass
-
-    def publish(self, msgs: List[Response], topic: str, schema_name=None):
-        self.publisher_thread_pool.submit(self._publish, msgs, topic, schema_name)
