--- pubsub/subscriber.py
@@ -1,7 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class Subscriber(ABC):
-    @abstractmethod
-    def subscribe(self):
-        pass
